refactor: extract config-settings-schema sub-schemas to restore maintainability floor (refs #3457)
