+++ baselines/crap.json
@@ -1,7 +1,7 @@
 {
   "$schema": ".agents/schemas/baselines/crap.schema.json",
   "kernelVersion": "0.1.0",
-  "generatedAt": "2026-06-03T02:03:38.718Z",
+  "generatedAt": "2026-06-03T02:15:55.742Z",
   "rollup": {
     "*": {
       "p50": 3,
